added mongo URI for Heroku

+++ financetrackerbackend/src/routes/chashflowApi.route.js
@@ -3,16 +3,16 @@ const cashflowApiRouter = express.Router();
 require('../mongo');
 const cashflowModel = require('../models/Cashflow.model');
 
-// cashflowApiRouter.use((req,res,next) =>{
+cashflowApiRouter.use((req,res,next) =>{
 
-//     if(req.session.user && req.session.user)
-//     next();
-//     else
-//     res.status(401).send('Forbidden! Maybe Login')
-// })
-// cashflowApiRouter.get('/', (req,res) => {
-//     res.send("You are logged in and can access dasboard" + req.session.user);
-// });
+    if(req.session.user && req.session.user)
+    next();
+    else
+    res.status(401).send('Forbidden! Maybe Login')
+})
+cashflowApiRouter.get('/', (req,res) => {
+    res.send("You are logged in and can access dasboard" + req.session.user);
+});
 
 cashflowApiRouter.post('/new', async (req, res) => {
     const newCashflow = await cashflowModel.create(req.body);
@@ -21,7 +21,7 @@ cashflowApiRouter.post('/new', async (req, res) => {
 });
 
 cashflowApiRouter.get('/view/:id', async (req, res) => {
-    const cashflowInfo = await cashflowModel.findOne({ _id: req.params.id }); //user_id: req.session.user.id} add  back in next
+    const cashflowInfo = await cashflowModel.findOne({ _id: req.params.id, user_id: req.session.user.id });
     console.log(cashflowInfo);
     res.send(cashflowInfo);
 });
@@ -34,7 +34,7 @@ cashflowApiRouter.put('/update/:id', async (req, res) => {
 
 cashflowApiRouter.delete('/:id', async (req, res) => {
     // req.body.user_id = req.session.user.id;
-    await cashflowModel.findOneAndDelete({ _id: req.params.id }); // user_id: req.session.user.id add  back in next
+    await cashflowModel.findOneAndDelete({ _id: req.params.id, user_id: req.session.user.id });
     res.status(200).json({ msg: 'Removed: ' + req.params.id })
 });
